requirements.txt
streamlit
google-generativeai
pandas
requests
fpdf2
openpyxl